Significant updates
Added a pile of new properties and methods to models, notably check_integrity methods to help secure database integrity.

The leaderboards page now links to decent filtered views as well to aid in diagnosing issues.

The leaderboard rebuilder and session addert have been reworked and sessions adds should now ostensibly (not even tested yet) update all ratings correctly. 

To make the monhtly email easier added a Copy button to the eladerboards page as well. This copies the leaderboards without links but otherwise as presented on the page, for convenient pasting into an email.

Certainly one of the best additions in this commit is the  ability to filter all list views by putting django query strings on the URL. WIll have to document the syntax some time.