Long overdue commit (Detail/List View options, Privacy support and more)
Lots of stuff going in here. Not all tested to satisfaction and still working on some. The bulk of what's been implemented here is that a whole pile of options have been added to the generic list and detail views including arudimentary cut at browsing records with the detail view (which I need to fix and is the next step planned but I need to upgrade to Django 2.0 to use SQL Window functions to do it right and so committing this  prior to that).

Privacy options are no supported but in beta I'd say, basic testing seems good but it was much harder to do cleanly than I expected and there is some duplicity (fields are hidden by the model on loading.

I expect to finish the browsing thing and some more tweaks and tests and commit again within the month.