Some tidy up of view options
Given Ih ad to make this half baked set of view options live recently I wanted quickly to create a view with indexes and a URL fetcher so I can share a nice view with someone.

Big part of this that still needs tidying is thatthe template now sees defaults and so when constructing URLs can avoid putting default values on the URL. That is all on the List View for  now. I have to implement this for the Detailview still.