Fixed puzzling bugs saving data ...
Not sure how these crept in. There's been a long pause in development as I sorted out Eclipse issues (it was unusable after an upgrade). Took me a month or more to fix that. So I wonder if I'd left development in a half baked state? And stupidly published it trying to fix the Django 2.1 issues in the last commit. It seems that saving data works now, but I have heavily instrumented the process while fixing this and it's high time we had some unit testing on the generic extensions. There are so many TODOs alas.