An intermediate check in ... that help fix the database after I broke it
I added a session with on player too many, the last one being on the list above. Wow that broke the leaderboards. So I've worked a tad at fixing form code to allow deletion of players from sessions. A long outstanding todo. Many little things as well, but that was the main drive and it's time to check in some code and secure progress. 

I'm in progress of fixing the team play submission code in the form. That remains broken for the moment. But it's working and running on the live site now so leaderboards are up again!