Significant form improvements to support Team mode sessions
A long outstanding fix, this a firstc ommit of provisionallytested form code and supporteing backend code, that now permits the entry of tem based sessions reliably and converting sessions  form one mode to the otehr (individual or team). This is all with an aim to fixing na old incorrectly netered Inkognito session and it tested fine on that. 

Before satisfying myself I'll need to to some rigours testing of all the code paths. But this passes provisional testing reliably submitting Team forms and allowing more flexible editing of sessions.