Got the DetaiView browser working again ...
Some minor little fixes as well, but main change introduced here is I got the dDetailView object browser working properly. Am in mid stream, there are still TODOs in the code. But committing this change to secure it.

Also added ordering format options to list dsiplay formates and object display formats and have yet to finish the implementation of that.