Rouded out Timezone support a little
Added timeonze fields against all DateTimeFields for now. This means we have a way of reording the ative timezone when datetime was saved which will empower sensible rendering of time's based on that knowledge. Also Locations won a TImezone that we can use at a late date, and I added default values to the Privacy field for Players (at long last) while reviewing the models. Along with some other tidy ups in the model space (adding verbose_names to evrything.