Moved the Trueskill Btea setting into the Game model
After studying trueskill more closely, Beta is used to model the luck component of a game. Tau is used to model the rate of forgetting when not playing in a sense - or better uncertainty about you skill during periods of absence (it's very poorly modelled in Trueskill as time is measured in plays not real time, a constant Tau variance between two game sessions - a heavy criticism of the current Trueskill model).