Added a Clear Session button to the Session Spy
I have a smal session spy, and wanted to test the Timezone updates on an empty session. So adde a Clear Session button to that. And also the timezone submitter now reloads the page if timezone changed, to ensure the view is rendered with the new timezone. That should probably be configurable, but will run with it for now, it introduces a slight screen flash and relaod on the first sruf in for a user. Might be able to reduce that with some research ... will see.