Select2 widget supplied ato context as HTML tested and facilitated
Tested on the League selector in the Header. Got a Select2 widget rendering and working well enough, but disabled for now. Was a proof of concept and I hope to generalise the Leaderboards options to take MultiSelect fields on the filters now (this widget enables that gracefully). That will take a little recoding effort, so parking code here for now as stable.