First phase of leaderboard view refactor
Not a complete refactor yet. Created a new options structure, fixed the HTML and JS to work. Still need to fix the fetch from GET/POST and then the leaderboard response to it and the title generation and some other tidy up. But wanting to commit the work so far as the refactor on the HTML and JS side is rather extreme.