Fixed that select2 placeholder rendering issue
Seems to be a bug in the selct2 widget, in that if it reders itself inside of a display: none element the plaeholder gets corrupted (truncated to about 6 chars). The fix was simply to trigger a chnage event when displaying the hidden advanced options, and it seems to render fine. Fixed a few other little niggles. but should now fix the backend to accept the options this submits (though first have to fix the submitter -
 function URLopts(element) to build a good submission from these controls.