Fixed a glaring bug
Form was not saving changes to model fields ....  Took some serious rethinking here and drilling into the get_form() and is_valid() methods to see what was going on. And took a little trickery to make the old object and the cleaned_data available to a pre_transaction and pre_save handler.