Score entering UI and Initial Testing Setp
Neither is complete. 

Updated the UI to accept scores in place of ranks for scoring games.
This predicates games being configured for scoring in the database.
The database fields to store them have existed for some while.

As the rank/score reconciliation prove complicated used this as a catalyst to add the initial bit of testing infrastructure. `manage.py test` now runs tests on session dicts and score reconciliation. 

It is not complete and this is an interim commit to secure work to date. 

Before going live need to add solid Request based testing for Scoring and non scoring games (submitting forms and checking that DB objects were properly created (that is ranks reflect scores and scores stored). That is the next step in testing.