Lots of debugging later
Leaves a fair bit of debug instrumentation in place, including the database configuration for testing. This commit simply secures a few work sessions that involved lots of debugging of session submissions, rank reconciliations and converting a session from team play to individual and back.

It is incomplete, with TODO: left, but basic mode conversion to and for is working without errors. Much form error handling has improved in the process, but again is incomplete.

This may serve well for publication to the staging server for testing there before tidying up some of the TODOs and publishing to live.

I hope to write some live server tests:

https://docs.djangoproject.com/en/4.1/topics/testing/tools/#liveservertestcase

before long too to cement form submission testing, that being the central (and most complex) interaction for data entry on the site.